+++ extensions/jira-time-tracking/CHANGELOG.md
@@ -1,8 +1,13 @@
 # Changelog
 
+## [Jira Server support] - 2024-04-20
+
+- Now supports both Jira Server and Jira Cloud instances.
+
 ## [Pagination support] - 2024-01-26
+
 - Support load all projects/issues with Jira pagination API.
 
 ## [Initial release] - 2022-04-26
 
-- Released first version of Jira Time Tracking 
+- Released first version of Jira Time Tracking